Added UserID field when deleting a contact

+++ LAMPAPI/delete.php
@@ -10,16 +10,17 @@ $inData = getRequestInfo();
 	else
 	{
         //Look for user in the database
-        $stmt = $conn->prepare("SELECT * from Contacts where Email like ? and ID=?");
+        $stmt = $conn->prepare("SELECT * from Contacts where Email like ? and ID=? and UserID=?");//added UserID when searching for the contact
 		$email = $inData["Email"];
-        $stmt->bind_param("si", $email, $inData["ID"]);
+        $stmt->bind_param("sii", $email, $inData["ID"], $inData["UserID"]);
         $stmt->execute();
         $result = $stmt->get_result();
 
         if(mysqli_num_rows($result) > 0){
 
-            $stmt = $conn->prepare("DELETE from Contacts where Email like ? and ID=?");
-            $stmt->bind_param("si", $email, $inData["ID"]);
+            $stmt = $conn->prepare("DELETE from Contacts where Email like ? and ID=? and UserID=?");
+			$email = $inData["Email"];
+            $stmt->bind_param("sii", $email, $inData["ID"], $inData["UserID"]);
             $stmt->execute();
 
         }else{
